+++ tests/test_visitors/test_ast/test_classes/test_base_classes/test_base_builtin_classes.py
@@ -11,6 +11,22 @@ enum_with_primitive1 = 'class First({0}, enum.Enum): ...'
 enum_with_primitive2 = 'class First({0}, Enum): ...'
 enum_with_primitive3 = 'class First({0}, EnumMeta): ...'
 enum_with_primitive4 = 'class First({0}, enum.EnumType): ...'
+enum_with_primitive5 = 'class First({0}, ReprEnum): ...'
+enum_with_primitive6 = 'class First({0}, enum.ReprEnum): ...'
+
+concrete_enum_with_primitive1 = 'class First({0}, enum.StrEnum): ...'
+concrete_enum_with_primitive2 = 'class First({0}, StrEnum): ...'
+concrete_enum_with_primitive3 = 'class First({0}, IntEnum): ...'
+concrete_enum_with_primitive4 = 'class First({0}, enum.IntEnum): ...'
+concrete_enum_with_primitive5 = 'class First({0}, IntFlag): ...'
+concrete_enum_with_primitive6 = 'class First({0}, enum.IntFlag): ...'
+
+enum_like_with_primitive1 = 'class First({0}, Choices): ...'
+enum_like_with_primitive2 = 'class First({0}, models.Choices): ...'
+enum_like_with_primitive3 = 'class First({0}, IntegerChoices): ...'
+enum_like_with_primitive4 = 'class First({0}, models.IntegerChoices): ...'
+enum_like_with_primitive5 = 'class First({0}, TextChoices): ...'
+enum_like_with_primitive6 = 'class First({0}, models.TextChoices): ...'
 
 
 @pytest.mark.parametrize(
@@ -70,6 +86,8 @@ def test_regular_subclass(
         enum_with_primitive2,
         enum_with_primitive3,
         enum_with_primitive4,
+        enum_with_primitive5,
+        enum_with_primitive6,
     ],
 )
 @pytest.mark.parametrize(
@@ -97,3 +115,83 @@ def test_builtin_subclass_with_enum(
     visitor.run()
 
     assert_errors(visitor, [])
+
+
+@pytest.mark.parametrize(
+    'template',
+    [
+        concrete_enum_with_primitive1,
+        concrete_enum_with_primitive2,
+        concrete_enum_with_primitive3,
+        concrete_enum_with_primitive4,
+        concrete_enum_with_primitive5,
+        concrete_enum_with_primitive6,
+    ],
+)
+@pytest.mark.parametrize(
+    'code',
+    [
+        'int',
+        'str',
+        'bool',
+        'list',
+        'dict',
+        'float',
+    ],
+)
+def test_builtin_subclass_with_concrete_enum(
+    assert_errors,
+    parse_ast_tree,
+    template,
+    code,
+    default_options,
+):
+    """Testing that it is not possible to subclass primitives and builtin Enum.
+
+    Builtin Enums are: `StrEnum`, `IntEnum` and `IntFlag`.
+    Already subclassed with primitives in standard library.
+    """
+    tree = parse_ast_tree(template.format(code))
+
+    visitor = WrongClassDefVisitor(default_options, tree=tree)
+    visitor.run()
+
+    assert_errors(visitor, [BuiltinSubclassViolation])
+
+
+@pytest.mark.parametrize(
+    'template',
+    [
+        enum_like_with_primitive1,
+        enum_like_with_primitive2,
+        enum_like_with_primitive3,
+        enum_like_with_primitive4,
+        enum_like_with_primitive5,
+        enum_like_with_primitive6,
+    ],
+)
+@pytest.mark.parametrize(
+    'code',
+    [
+        'int',
+        'str',
+        'bool',
+        'list',
+        'dict',
+        'float',
+    ],
+)
+def test_builtin_subclass_with_enum_like(
+    assert_errors,
+    parse_ast_tree,
+    template,
+    code,
+    default_options,
+):
+    """Testing that it is not possible to subclass builtins and enum-like."""
+    tree = parse_ast_tree(template.format(code))
+
+    visitor = WrongClassDefVisitor(default_options, tree=tree)
+    visitor.run()
+
+    assert_errors(visitor, [BuiltinSubclassViolation])